requirements.txt
fastapi
uvicorn
weasyprint
jinja2
pydantic
python-dotenv
